+++ ChartJs.Blazor/ChartJS/PolarAreaChart/PolarAreaDataset.cs
@@ -36,6 +36,6 @@ namespace ChartJs.Blazor.ChartJS.PolarAreaChart
         /// </summary>
         public List<int> HoverBorderWidth { get; set; }
 
-        public List<int> Data { get; set; }
+        public List<double> Data { get; set; }
     }
 }
